Refactor SettingsScreen styles: move to new directory structure, update import paths, and enhance styling consistency
