+++ src/translation/locales/en.json
@@ -10,7 +10,7 @@
   "toastSaveMessage2": "Group settings have been saved",
   "deanGroup": "Dean's Group",
   "labGroup": "Laboratory",
-  "compGroup": "Computer Lab",
+  "compGroup": "Computer Laboratory",
   "projGroup": "Project",
   "freeHoursText": "Display free periods",
   "language": "Language",
